digital lights control: rename drivers/lights_hal.{c,h} to drivers/lights_io.{c,h}
